factor out rendering
This is the first version that pulls out the model, texture, etc. loading and puts it in a separate class. the approach I'm going with is inheritance with some functions to override in the derived class. Why am I doing this even though I hate inheritance with a passion? I'm not sure, just lazy, I'll fix it later.

new structure:
- samples dir for things to render
- Vulk dir for the renderer

Also add verbose warnings